python/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "robot-configurator"
version = "0.1.0"
description = "LLM-driven workspace configurator and motion planner for the Franka Emika robot arm"
readme = "README.md"
requires-python = ">=3.8"
license = { file = "LICENSE" }
authors = [
    { name = "Sandro Gabl" }
]
dependencies = [
    "openai>=1.0",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.4",
]

[tool.hatch.build.targets.wheel]
packages = ["src/robot_configurator"]

# ── Ruff (linter + formatter) ──────────────────────────────────────────────
[tool.ruff]
line-length = 100
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B"]
ignore = ["E501"]  # line-length enforced separately

# ── Pytest ─────────────────────────────────────────────────────────────────
[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"

# ── Coverage ───────────────────────────────────────────────────────────────
[tool.coverage.run]
source = ["src/robot_configurator"]
omit = ["tests/*"]
